requirements.txt
click==8.1.7
expiring_dict==1.1.0
Flask==3.0.3
flask_jwt_extended==4.6.0
flask_script==2.0.6
loguru==0.7.2
python_bcrypt==0.3.2
